+++ trunk/black_scholes/objects/stock.c
@@ -8,7 +8,7 @@
 
 using namespace std;
 
-int calculate_stock( obj_loc_t obj_loc, int id )
+int calculate_stock( obj_loc_t obj_loc, long id )
 {
     stock_vv = stock_v * stock_v;
